# Fix Android startup perf regression