Patch `supportedLocales` from  `@formatjs/intl-pluralrules` to only include the ones we actually use. Since this is easy to mess up, I also added a lint rule that enforces we don't add new locales without also tweaking that patch.

I'm not sure why the `@formatjs/intl-pluralrules` polyfill is so slow but I didn't have time to get into it. I filed https://github.com/formatjs/formatjs/issues/4463.

## Test Plan

Disable the `console`-removing Babel plugin in the Babel config.

Before (5 runs):

```
Time to first paint: 2675 ms
Time to first paint: 2926 ms
Time to first paint: 2762 ms
Time to first paint: 2928 ms
Time to first paint: 2853 ms
```

After (5 runs):

```
Time to first paint: 1905 ms
Time to first paint: 1921 ms
Time to first paint: 1926 ms
Time to first paint: 1929 ms
Time to first paint: 1901 ms
```

I haven't verified what this part of the `Intl` polyfill _actually_ does but it seems safe to assume that it's OK to exclude the locales we don't use. 

Switching languages seems to still work.