@haileyok 2024-06-18
Three runs on a physical device w/ patch:

```
Time to first paint: 1235 ms
Time to first paint: 1166 ms
Time to first paint: 1207 ms
```

Three without the patch:

```
Time to first paint: 1888 ms
Time to first paint: 1893 ms
Time to first paint: 1924 ms
```

Also tested switching between languages, doesn't seem to affect anything there. Seems solid, great investigation! 🙌